Remove duplicate reloc enums
