+++ spimdisasm/mips/__init__.py
@@ -13,4 +13,3 @@ from . import FilesHandlers
 from .InstructionConfig import InstructionConfig
 from .MipsFileBase import FileBase, createEmptyFile
 from .MipsFileSplits import FileSplits
-from .MipsRelocTypes import RelocTypes
