+++ spimdisasm/mips/symbols/MipsSymbolFunction.py
@@ -8,7 +8,6 @@ from __future__ import annotations
 import rabbitizer
 
 from ... import common
-from ... import elf32
 
 from . import SymbolText, analysis
 
@@ -377,15 +376,15 @@ class SymbolFunction(SymbolText):
         return was_updated
 
     _percentRel = {
-        elf32.Elf32Relocs.MIPS_HI16.value:      f"%hi",
-        elf32.Elf32Relocs.MIPS_LO16.value:      f"%lo",
-        elf32.Elf32Relocs.MIPS_GPREL16.value:   f"%gp_rel",
-        elf32.Elf32Relocs.MIPS_GOT16.value:     f"%got",
-        elf32.Elf32Relocs.MIPS_CALL16.value:    f"%call16",
-        elf32.Elf32Relocs.MIPS_GOT_HI16.value:  f"%got_hi",
-        elf32.Elf32Relocs.MIPS_GOT_LO16.value:  f"%got_lo",
-        elf32.Elf32Relocs.MIPS_CALL_HI16.value: f"%call_hi",
-        elf32.Elf32Relocs.MIPS_CALL_LO16.value: f"%call_lo",
+        common.RelocTypes.MIPS_HI16.value:      f"%hi",
+        common.RelocTypes.MIPS_LO16.value:      f"%lo",
+        common.RelocTypes.MIPS_GPREL16.value:   f"%gp_rel",
+        common.RelocTypes.MIPS_GOT16.value:     f"%got",
+        common.RelocTypes.MIPS_CALL16.value:    f"%call16",
+        common.RelocTypes.MIPS_GOT_HI16.value:  f"%got_hi",
+        common.RelocTypes.MIPS_GOT_LO16.value:  f"%got_lo",
+        common.RelocTypes.MIPS_CALL_HI16.value: f"%call_hi",
+        common.RelocTypes.MIPS_CALL_LO16.value: f"%call_lo",
     }
 
     def generateHiLoStr(self, instr: rabbitizer.Instruction, instrOffset: int, symName: str, symbol: common.ContextSymbol|None, relocInfo: common.ContextRelocInfo|None=None, gotHiLo: bool=False) -> str:
